Changed how users are added to spaces and added Staging space
Each space now has a .csv listing users to add. This is to help
generalise the deployment process for both spaces for individuals and
group spaces such as staging.

`ci/tasks/create-users.yml` will replace $SPACE and $ORG with the
necessary values and output a csv file to be used by cf cli
`create-users-from-file`.

with @vixus0