Single task to apply network policies
We were running into issues with too many containers on our Concourse
workers. We probably don't need to have a separate `put` for each and
every network policy. Instead, we can use `paas/network-policies.yml` as
the canonical source for network policies and apply them all in one go.

Something to think about is that this won't remove already existing
policies but that could be added later. Also I'm not using Andy's cf-cli
plugin because it needs Go to install.