Terraform for Cloudfront
I've created an `environment` Terraform module that holds configuration
for a single Pay environment (for example, staging).

Inside `environment` we setup a Cloudfront distribution to forward
traffic to the PaaS and some Route53 records to setup the environment's
subdomain and an ALIAS record to the Cloudfront distribution.

Cloudfront can only use ACM certificates that were provisioned in
us-east-1, so we need to provide an additional aliased `aws` provider to
the `environment` module that can be used for provisioning that cert.