Use Terraform to provision Pay PaaS environment
Terraform is a better option for provisioning the "infrastructure" in
PaaS. It also means we can properly automate provisioning the
environment using Concourse.

This makes use of the [Cloud Foundry provider for
Terraform](https://github.com/cloudfoundry-community/terraform-provider-cf)
which needs to be downloaded and installed in the Terraform plugins
directory.

We've added the following new Terraform modules:

- `paas`: Defines the placeholder apps, network policies and the app catalog service
- `paas-postgres`: Defines the RDS service instances
- `paas-stub-postgres`: Defines a stub postgres service for dev envs
- `paas-stub-sqs`: Defines a stub SQS service for dev envs

For the `stub` modules, the `paas` module needs to be provisioned first:

```
terraform init
terraform apply -target=module.paas
terraform apply
```