Use Terraform to provision Pay PaaS environment
Terraform is a better option for provisioning the "infrastructure" in
PaaS. It also means we can properly automate provisioning the
environment using Concourse.

This makes use of the [Cloud Foundry provider for
Terraform](https://github.com/cloudfoundry-community/terraform-provider-cf)
which needs to be downloaded and installed in the Terraform plugins
directory.