Add pre-commit config
We should be running some validations when we modify the Concourse
pipelines or Terraform config.

This uses [pre-commit](https://pre-commit.com) to run the validations
provided by `fly` and `terraform`.

To try it out manually on OSX:

```
brew bundle install
pre-commit run -a
```

To configure `pre-commit` to run every time you `git commit`:

```
pre-commit install
```

I've also added a Travis CI configuration that will run the pre-commit
checks when a PR is raised.