Add script for provisioning a production PaaS space
We need some way of defining a production(-like) PaaS space. Terraform
is probably too heavyweight as a solution for the time being. I've
written a bash script that does the following:

- Create a normal app space
- Create a separate CDE space if required
- Create placeholder buildpack apps
- Add network policies between all the apps
- Sets up deployer user permissions

Todo:

- Update network policies and script for CDE space
- Add SpaceAuditor permissions for Pay devs
- Create services to hold credentials

For the time being, I expect the script to be manually run by an OrgManager
user to provision the space.