Use terraform deployment per environment
We want the distinction between separate environments (staging, prod) to
be more explicit, so that people don't accidentally apply changes to
multiple environments when running terraform.

Now, to apply the staging terraform, you need to be in the `staging`
directory before running `terraform apply`.

The `account` directory holds terraform that should be shared across the
AWS account.