PP-6030 Correct env vars for staging
In the new individual manifests we are no longer using the format
`http://((app_route)):8080` so the urls in the config need to be
updated. I've tried manually changing the env vars for card-frontend to
those included in here and the smoke tests pass.

We may have some inconsistency as we move to individual manifests and
buildpacks whether the manfiest expects the protocol prefix or not, if
things break we can fix it up.