PP-6031 Switch java apps to buildpacks
Update java apps to deploy with buildpack using the manifest in the
latest git pre-release. Update the staging envs file to provide value
necessary to get the apps running. These values may not be sufficient to
get the smoke tests passing but we are dealing with env vars in future
stories.