PP-6031 Correct archive locations and a few envs
Need to specify the jar location whilst keeping path a directory for the
Node apps; the path in the manifest assume the push command is made from
the directory above `target` however on concourse we're actualy in the directory
above artefact e.g path must be `./artefact/target/pay-*-allinone.jar`.

Probably a nicer way of doing this but we can iterate.