Ignore changes to health_check_type
We think the provider should be ignoring these anyway, but it looks like
it wants to change them all back to port healthchecks. Possibly a bug in
the cloudfoundry terraform provider.

We're a bit worried we might end up playing whack-a-mole ignoring things
like this, but it seems like `health_check_type` is the only one that's
a problem at the moment. I'll raise an issue with the terraform provider
to clarify whether this is working properly or not.