PP-6146: Ensure Notifications app is pushed as a Docker app
This was being created as a buildpack-based app, which causes
initial deployment of the Docker container to fail since
app types can not be changed without deleting first. Since
the notifications app container is specified using manifest.yml,
Terraform will create an empty docker app and ignore the image
in subsequent runs.