PP-6444 Add e2e tests to pr pipeline
Restructured the pr pipeline whilst adding the e2e tests. Split testing
of java apps into three jobs, unit, integration and end-to-end. This
means their status can be reported to Github individually, they can run
in parallel and re-run independently.

Node apps are tested in a single job because installing and compiling dependencies
is costly and vary between dev and production environments (e.g. whether
Cypress should be included), thus the benefit of running in parallel is
lost.

Imported end-to-end scripts from pay-scripts since only a subset were
required and they have been updated to run with this pipeline. We may
wish to revisit this. We have made some small improvements to this:
 - use docker health check rather than old ready.sh
 - use selenium standalone rather than hub + chrome node