PP-6686 Tag consumer master pacts with master
Upon merge commit of a consumer find the previous commit that was not a
merge. This should be the head commit of the PR that was tested and
which was just merged. The pact broker will have contracts for this version
of the consumer. Use curl to tag them with "master".

This avoids generating the pacts a second time for node apps which
would require an npm install which is costly.

This was the simplest solution at the time and we may want to revist it.