PP-6704 Run provider pact validation on master merge
When a provider is merged into master run pact validation using the
commit one before the merge commit which will be the commit from the
head of the PR. The pact broker will know about this commit since it
will have been tested during the PR build leading up to this merge. If the
pacts with its consumers are still valid then exit. If not then rerun
the validation for the consumer. If the rerun fails then manual
intervention will be necessary as it would be today with Jenkins CI.