BAU Fix adminusers db migration script
When running db migrations we use `source` to modify & execute the start command directly
from the app's container to future proof against buildpack changes and the command
changing.

The app's start command is run with `exec` which replaces the process
that called it with a new process. This makes sense when you're starting the
app and want the app process to be pid 1.

When we're running migrations we want the command we've provided to `cf run-task`
to remain as pid 1 especially for adminusers where we effectively
run `initial migration && all other migrations`. Without this change the
initial migration runs and the container exits once its completed
and the next migration does not happen.