Update deploy pipeline using new set-pipeline job step
We no longer need to use the update pipeline resource now
Concourse has a native method to update piplines on master merge.

I have added "git" resource to track changes on master to the
deploy.yaml pipeline. This ensures the update is only triggered
when the pipline is changed, rather than any file in the omnibus
repo.