PP-6832 Stop containers used for e2e
We effectively run `docker system prune -af --volumes` when the e2e task
exists via a trap. However we're not stopping the containers used in the
e2e tests so they're not pruned when that is run. This leaves the
containers and images handing around. If we stop the containers before
running `docker system prune` it should clear them all out and free up
more disk quicker.