PP-7013 Add main pipeline
Initial work on adding main pipeline to build and create github
releases destined for paas. For the time being Jenkins will continue to run the tests on
the main branch and this pipeline will not; we may want to run the pact
tests because we're using separate brokers and `can-i-deploy` checks will be much
more complicated if we don't. Later work will migrate the testing from
Jenkins to this pipeline.

Changes included here:
- Build connector when a commit to main branch is made
- Create and upload github pre-release tagged `paas_release-n`
- Remove building of connector release from existing `deploy` pipeline
- Trigger the deployment of connector into staging via a release tagged
  with `paas_release_test-n` which is created via the
  `pay-dev/test-deploy` pipeline.