PP-7025 Use separate hosted zones for sub domain delegation
Currently the staging account has the domain registration for
`gdspay.uk`. To make the terraform consistent for any environment
(test, staging etc) create a hosted zone in which the A records for that
environments' apps are created in. Delegate the subdomain to the new  hosted
zone from the hosted zone in staging which owns the `gdspay.uk`
registration; this is to be done manually for now by adding the
necessary name server record pointing to the name servers of the new
hosted zone.

Remove subdomain interpolation in favour of using `var.domain_name`
passed in from the environment module, e.g. `staging.gdspay.uk`..

Co-authored-by: dan worth <dan.worth@digital.cabinet-office.gov.uk>