PP-7026 Pass the consumer tag as a variable into pact tasks
Previously we were only concerned with tagging and verifying pacts with
the tag `master`. Now we are deploying into `test` environment we need
to tag and verify with the tag `test`. This change passes the tag to use
for these actions into the task files.

Substitute `main` for `master`. Next commit will rename files
accordingly, kept separate to help review the refactoring done herein.