PP-7022 Fix can-i-deploy check
Switch to `json` output which is easier to parse. The preivous logic was
not correctly identifying providers that required validation when they
had not yet been validated against the version of the contract in the
request. This uses a new container which adds `jq` to the
`pactfoundation/pact-cli` image we were already using.