PP-7021 Create paas test releases
After a pre-release with the tag format `paas_release-n` has been
deployed and the smoke tests have passed, create a pre-release with the tag
`paas_release_test-n` which can then be deployed into staging.

This is just for Adminuers and will apply the same to the other apps if
we agree on this approach.