PP-7027 Remove smoke test user setup
In the early days of the PaaS migration we used a bit of the `pay local`
cli to create a user to be used for the sandbox smoke tests. Now we are
using snapshots of RDS instances from our existing environments e.g.
`test-12` we want to use the existing smoke test services and accounts.

Therefore remove the concourse job that would create a new smoke test
user and add a simple script to help set the necessary values.