PP-7121 Remove pact validation prior to deployment
This is temporary. Whilst we're not running the pact tests on a merge to
the main branch on concourse the concourse pact
broker does not have a reference to the main branch merge commit of the
consumer or provider- only that of the PR head commit which was tested on the pr
pipeline.

When we deploy a consumer we need to tag it's pacts with the environment
which is used in `can-i-deploy` checks. When we get the git release to
deploy we only know the main branch merge commit it is associated with.
The pact broker does not have any reference to this so we cannot make a
call to add the environment tag to that unknown version of the consumer.
Similarly this impacts providers because although the PR commit sha will
have been recorded as validated during the PR build, the merge commit sha will
not and the validation would have to be needlessly run a again on deployment.

The solution is to run the pact broker validation and push contracts to the
pact broker when building the main branch on `pay-dev/main` which will be done
at a later date. Once that is done we can re-instate tagging consumer pacts and
provider versions upon deployment which will permit `can-i-deploy`
checks.

The work to reinstate this is captured in jira epic PP-7204
https://payments-platform.atlassian.net/browse/PP-7204