PP-7502 Remove unused ci task files
Having removed the unused concourse pipelines this commit removes the
now unused task files. Some of these files might be useful to reference
when we come to build the pipelines to build and deploy fargate applications to
AWS. They are being removed from the main branch since they're no longer
used and to avoid any confusion.