PP-7652 Remove all tf, docs, paas things
Removes all terraform, including some that corresponds to infrastructure
that hasn't been removed yet (mainly IAM roles for access to defunct accounts)
but I figure we can click ops them into oblivion when needed.

Removes the entire omnibus docs pipeline as that does not seem
to be a thing we use these days.

Removes entire paas dir as not needed.

I guess this purge might be too extreme, but as the old saying goes,
"better to delete too much than too little". Or "it's always in git"