PP-7657 - Concourse toolbox ECR test pipeline
Triggers on a git commit matching the "alpha_release-*" pattern on the Toolbox repository.

Pulls the image matching that commit SHA from Dockerhub, saves the image locally, then pushes to ECR.

We're using a custom registry-image resource type. This is based on Richard Baker's work done on a similar Notify pipeline. Support for the registry_id parameter is not yet available in the stable release, so we're fetching it from the main github branch for the moment. This will allow our Concourse user to assume the correct role and push to the ECR registry.