PP-7794 Split toolbox test pipeline in to build and deploy
Splits existing toolbox test pipeline into two:

First pipeline triggers on git release, then takes an image from
dockerhub and pushes it to test ecr.
The only change here is that on pushing to test ecr, we add a tag
to the image of the form 123-release. This is in order to trigger
the second pipeline

The second pipeline is triggered by a push to test ecr of an image
with tag of the form `123-release`. THis pipeline has two dummy
jobs to deploy and smoke test this new image (these were previously
tasks, but it makes things clearer to have them as jobs). It then
pushes the image to staging ecr. No tags are added at this stage,
since the staging pipeline only needs a tag of form above (and this
also allows us to refer builds back to originating code change).