PP-7834: Get git sha for a release tag
In order to tag a pact with something (eg "test-fargate"), we need the version
which is the git commit sha of the release tag.

In line with the agreed practice of writing scripts in node, we have decided to
make use of a github node api. To this end we have added the `octokit/rest` npm
library to our node-aws-sdk-runner and renamed it to node-runner.