PP-7686 Add deployment to staging toolbox pipeline
Add deploy step to toolbox staging pipeline. I have also split
the pipeline up similarly to test pipeline, and introduced
job groups.
As part of this I have removed the task that passed through
additional tags from the staging image through to the prod
one. This is because we now don't pass any such tags from
upstream. If we need to e.g. retrieve the git sha for code
corresponding to an image we have a task that can do that.