PP-7881 Node.js to run a db migration on Fargate from Concourse
Follows the same principles as our existing ruby script run on Jenkins.
Run the current active task defintiion for the service (e.g. Ledger)
whilst overriding the `RUN_MIGRATION` and `RUN_APP` to `true` and
`false` respectively.

If the image version used as the input to the job does not match the
image version of the app currently running then print a warning and
exit.

Update the test deploy pipeline with a ledger-db-migration job.