PP-7877 Trigger toolbox deploy on new telegraf
We plan to use Toolbox deployments as a way of ensuring
new versions of sidecars are not totally broken.
This commit enbables this by triggering a new deploy
of toolbox when a new version of telegraf appears in
test ECR. The job then deploys this new version (alongside
version of toolbox specified by toolbox ecr resource)
I have split into an inline task so as not to have any
confusing logic in the deploy-app task.