PP-7889 Notify success and failure of toolbox test deployments
Couple of funny looking things:
The message is partly comprised of a bit written to a file and then
loaded to a variable. This allows us to move the task that creates
the messages to a task file, parameterise etc.
The other bit of the message (the build information) is stuck on the
end in the `on_{success,failure}` handlers. This is because Concourse
only exposes build information in resources
https://stackoverflow.com/a/40939590
so if this were appended to the message in the create messages task,
all the env vars would evaluate to an empty string.