PP-7873 Add pipeline to update the smoke tests
The deploy-smoke-tests take a release of pay-smoke-tests that
is built by github actions. There is a manually triggered job to update
the canaries in test, staging and production where the pay-smoke-tests
release must have been deployed to the previous environment before
progressing. This commit does not include any automatic running of
canaries post deploy.