PP-7887: Check Canary is in a startable state
Use an interval to check the canary status. If it's in status STOPPED then we can attempt to start it ourselves. If it then subsequently fails due to a conflict (something else has triggered a run) then we can just use that run, as we know it will have begun after our deploy finished.