PP-7908 Check release versions before deployment
In our current Jenkins pipelines we have a check that
stops devs accidentally deploying an old version of an
app. This `check-release-versions` task achieves same thing
on Concourse - it checks the version to be deployed of each
container in a task definition, and makes sure it is the
same as or newer than currently deployed version.

There is no mechanism to override this at the moment -
if override is needed devs will need to deploy the thing
manually from laptop.