PP-8162: Deploy carbon-relay to test env
Deploy carbon-relay to the test environment. A notable change in this commit
would be for the the `ci/scripts/wait-for-deploy.js` file which was amended to
make it work with carbon-relay. As carbon-relay has no `APPLICATION_IMAGE_TAG`,
`NGINX_IMAGE_TAG`, or `TELEGRAF_IMAGE_TAG` these parameters are now made
optional.

This commit does not include the creation of notification snippets, the
checking that we aren't deploying an older version or pushing the image to
staging ECR. These will be done in a future PR.