PP-8162: Deployment to prod should be a manual process
At the moment we are of the opinion that carbon-relay shouldn’t be continuously
deployed into staging/production because we don’t have smoke tests that cover
whether or not app metrics are being sent. That said, we don’t send app metrics
from test-12 so it has to go into at least staging before one could make an
assessment. Therefore we are leaving `trigger: true` on staging, whilst
deployment to prod will have to be manually triggered.