PP-8105 Add Notifications Docker to ECR Job
Notifications is being built by Jenkins CI which pushes it into
Dockerhub as per the other apps. This Job triggers on releases in
pay-notifications github repo, pulls the image from Dockerhub and pushes
into ECR. This is the same pattern we have for the other apps built by
Jenkins CI until we move the build somewhere else.