PP-8105 Add Deploy Notifications Job
Adds a job to deploy notifications app, run smoke tests and then push
the image into staging ecr. This follows the existing pattern with a
slight change in the deploy task. Since the notifications app consists
of a single container it doesn't lend itself to reusing existing deploy
yaml files so instead it has been in-lined. Since this is specific to
the notifications app and its a simple bit of bash there it isn't worth
putting this into a yaml file for re-use at this time.