PP-8729: Fix github release resource
Github recently changed their release formatting behind the scenes (see https://github.com/concourse/github-release-resource/pull/107). The fix has been pushed for the Concourse github-release-resource, but this is not available in the version of Concourse that BigConcourse is currently running (v7.2.0). Given it may be 'a while' before BigConcourse reaches 7.5.0+, we can instead pin to the version of the resource repo that contains the fix, and specify this as a resource_type.

The naming of the resource type is a bit clumsy but I want it to be clear that this is a version of the resource, not the bleeding edge of the repo. Suggestions for a better name welcome.