BAU Stop running Cypress tests on PR builds
Cypress tests are now run by a GitHub action on pay-frontend and
pay-selfservice and it is mandatory for this check to pass before a PR
can be merged.

Therefore it is no longer necessary to run them as part of the concourse
PR pipeline. We still need to run the unit tests as the output of these
is used for the pact verification job, which hasn't been moved to GitHub
actions yet.