PP-9096: Clarify the job formerly known as 'pact-provider-verification'
Connector and ledger PR builds contain a 'pact-provider-verification' job, in
addition to a 'pact-provider-test' job which other app PR builds have. This is
because connetor and ledger are both consumers and providers in the pact world.
The 'pact-provider-verification' job actually runs pact provider tests where
they are the consumer. This PR gives clarity to this purpose. We also set the
context as 'app-as-consumer pact verification' which is used by the GitHub PR
as an additional check.

Naturally, we could clarify the 'pact-provider-test' job and rename it to
'app-as-provider pact verification' but that will come in another PR.