PP-9236: Workflow for running a provider contract test
The pro of having this workflow is that the maven command to run this test is
the same across all our providers.

The cons:
* The providers must be in Java
* The provider project must have a pom.xml which defines a `runContractTests`
profile

I've tested that we don't need to pass `-Dpact.provider.version=${commit}` to
the `mvn test -DrunContractTests` command as the pact library will pull out the
latest sha of the provider when publishing the result.