BAU Run Pact tag task before pushing to ECR
For all pipelines on deploy-to-test and for some pipelines on
deploy-to-staging, the `push-x-to-production-ecr` needs the `x-pact-tag`
job to have completed first.

However, for a few apps on the deploy-to-staging pipeline, the
`push-x-to-production-ecr` requires the smoke test task to have
completed. This means that we might push to the production ECR without
tagging the Pact for staging if Concourse gets stuck, which might cause
confusion if we later run across Pact errors.

Fix this so the ordering of tasks across all pipelines is uniform.