BAU: Exclude version check from test-perf-1 DB migrations
We ran into problems with this as the image tagged 2856-perf-db had not been deployed as an app release (which was still on 2855-perf).

We don't carry out version checks for test-perf-1 deployments elsewhere, as it's low risk given the perf environment should be up to date with production deployments.

The alternative would be some more complex changes to our application deployment script, or an additional app deployment job just for the *-perf-db tagged images (which would probably cause different problems for this check anyway).