PP-12537: Concourse pipeline to build pay-java-lambdas JARs (#1159)
* PP-12537: Pipeline to build pay-java-lambdas

The build works by getting a list of submodules to build. This list is got by looking at the changes from the last commit in the main branch. Only those submodules are built, which is why although `mvn versions:set -DnewVersion="$PACKAGE_VERSION"` changes the versions for all the submodules, it doesn't matter.