PP-12421: Define deploy-to-staging with pkl
This is the pkl version of the deploy-to-staging pipeline. The intent is for this to have minimal change to the concourse pipeline, however since the original yaml was not internally consistent between apps there are some minor changes where the pkl definition forces that consistency. We also add reporting metrics.