PP-13785 Refactor
Remove abstractions and for loop by account. pay-dev deploys to dev and
test only, and these accounts have substantially different setup (dev
only has has account level pipeline) and whether we want to destroy or
run E2E tests is different between these environments. We can abstract and
refactor if we think it makes sense after modifying this pipeline to
meet our design of testing in test and keeping development off of the
path to production.