[WIP] PP-14175: Fix local e2e container healthchecks
The containers that use the reverse_proxy (frontend and selfservice) are failing the healthchecks when run locally, as the healthchecks are configured using localhost instead of the proxy alias URL.

Changing the healthcheck URLs to use the alias, e.g. frontend.internal.pymnt.localdomain:9200, allows the healthcheck to pass locally.